(libfreefare-desfire) Replace the test_mifare_desfire test-suite's tests with a single test_mifare_desfire() test.
Tests-suites which act on real cards are a PITA to write and maintain because they introduce a log of redundancy.  A single tests that checks everything looks easier to write and maintain.  This code is inspired by some Mifare DESFire test-suite, but only parts currently implemented are tested for now.  A good deal of rewriting is required to make this code more bearable, and running the test-suite often fails randomly... there is a problem somewhere I guess (and I am not sure it is a hardware problem...) but it is development code right?

Note to myself: do not call mifare_desfire_change_key() with uninitialized (free'd) new_key data.