(libfreefare-desfire) Report write_data() returning 0xAF as an error condition.
With the current code, sending plain-text data while we are supposed to send
enciphered data result in write_data() returning with the 0xAF PICC error code.
This is actually an error condition.  For now it is due to a limitation in
libfreefare which should be fixed in a few hours / days (1st priority in my
TODO list), and this error should not be reproducible.  I prefer however to
detect and catch this now instead of having an error triggered when the next
command is executed in the test suite.